# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TEST API FISCAL IO",
        "description": "Explore a API!  Nossa plataforma API oferece uma gama abrangente de endpoints para acessar e interagir com os dados essenciais. Descubra recursos robustos e serviços projetados para simplificar e enriquecer sua experiência na obtenção e gerenciamento de informações acadêmicas da mais alta qualidade.\r\n    [http://swagger.io](http://swagger.io) or on\r\n    [irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "contact": {
            "name": "Time",
            "email": "jhonnatasimoes@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080/",
            "description": "API Url Server"
        }
    ],
    "paths": {
        "/number-in-words/{number}": {
            "get": {
                "tags": [
                    "NumberInWords"
                ],
                "summary": "Get a NumberInWords.",
                "description": "Write the number in full, for example: One, two..",
                "operationId": "3517960bf4dc8399c1fb4a4f11d554d1",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Number for get name",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}